refactor: remove types in separate files
