+++ src/components/Checkbox/Checkbox.type.ts
@@ -1,0 +1,4 @@
+export type CheckboxProps = {
+  checked: boolean;
+  onChange: (checked: boolean) => void;
+};
