+++ src/components/CoinsList/CoinsList.type.ts
@@ -1,0 +1,5 @@
+import { Coin } from '../../store/api/coinsApi.type';
+
+export type CoinTableProps = {
+  coinsList: Coin[];
+};
