+++ src/components/ErrorBoundary/ErrorBoundary.type.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from 'react';
+
+export type Props = {
+  children?: ReactNode;
+};
+export type State = {
+  hasError: boolean;
+};
