+++ src/components/Loader/Loader.type.ts
@@ -1,0 +1,3 @@
+export type LoaderProps = {
+  role?: string;
+};
