+++ src/components/Pagination/Pagination.type.ts
@@ -1,0 +1,5 @@
+export type PaginationProps = {
+  currentPage: number;
+  totalPages: number;
+  onPageChange: (page: number) => void;
+};
