+++ src/components/Search/Search.type.ts
@@ -1,0 +1,4 @@
+export type SearchProps = {
+  updatedCoinsList: (searchRequest: string) => void;
+  isDisabled: boolean;
+};
