+++ src/hooks/reduxTypedHooks.ts
@@ -1,5 +1,5 @@
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import type { RootState, AppDispatch } from '../redux/store';
+import type { RootState, AppDispatch } from '../store/store';
 
 export const useAppDispatch: () => AppDispatch = useDispatch;
