+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/App';
 import { Provider } from 'react-redux';
-import { store } from './redux';
+import { store } from './store/index.ts';
 import ErrorBoundary from './components/ErrorBoundary';
 import { ThemeProvider } from './context/ThemeProvider.tsx';
 import './index.scss';
