--- src/redux/index.ts
@@ -1,2 +1,0 @@
-export * from './store';
-export * from './coinsApi';
