+++ src/store/index.ts
@@ -1,0 +1,2 @@
+export * from './store';
+export * from './api/coinsApi';
